conv.catch
In JABBAv.1.1.R, line 79

conv.catch = as.numeric(rbind(matrix(rep(NA,(styr.I-1)*n.catches),styr.I-1,n.catches),as.matrix(catch[,-1])))

this statement added extra NAs for catch matrix where in many cases may be not necessary.

Our data sets of catch and cpue series are with the same length and with above line, it doesn't work. We simply changed it as below and it works:

conv.catch = as.matrix(catch[,-1])